restructure directories to resemble other threlte packages
