+++ package.json
@@ -46,7 +46,7 @@
 		"svelte": "^4.1.2",
 		"svelte-check": "^3.4.6",
 		"tailwindcss": "^3.3.3",
-		"three": "^0.155.0",
+		"three": "^0.153.0",
 		"three-inspect": "^0.3.4",
 		"tslib": "^2.6.1",
 		"typescript": "^5.1.6",
