+++ src/lib/components/controller.svelte
@@ -9,9 +9,9 @@
 import { onMount, onDestroy } from 'svelte'
 import { T, useThrelte, createRawEventDispatcher } from '@threlte/core'
 import { XRControllerModelFactory } from 'three/examples/jsm/webxr/XRControllerModelFactory'
-import { fire } from '$lib/events'
+import { fire } from '$lib/internal/events'
 import type { XRController, XRControllerEvent } from '$lib/types'
-import { isHandTracking, activeTeleportController, pendingTeleportDestination } from '$lib/stores'
+import { isHandTracking, activeTeleportController, pendingTeleportDestination } from '$lib/internal/stores'
 import { left, right, gaze } from '$lib/hooks/use-controller'
 import ShortRay from '$lib/components/ray-short.svelte'
