+++ src/lib/components/hand.svelte
@@ -4,7 +4,7 @@ import { onMount, onDestroy } from 'svelte'
 import { T, useThrelte, createRawEventDispatcher, useFrame } from '@threlte/core'
 import { XRHandModelFactory } from 'three/examples/jsm/webxr/XRHandModelFactory'
 import type { XRHandEvent } from '$lib/types'
-import { fire } from '$lib/events'
+import { fire } from '$lib/internal/events'
 import { left, right } from '$lib/hooks/use-hand'
 
 const handModelFactory = new XRHandModelFactory()
