+++ src/lib/components/hands.svelte
