+++ src/lib/hooks/use-hit-test.ts
@@ -1,6 +1,6 @@
 import * as THREE from 'three'
 import { onDestroy } from 'svelte'
-import { session } from '$lib/stores'
+import { session } from '$lib/internal/stores'
 import type { HitTestCallback } from '$lib/types'
 import { useThrelte, useFrame } from '@threlte/core'
