+++ src/lib/hooks/use-teleport.ts
@@ -1,6 +1,6 @@
 import * as THREE from 'three'
 import { useThrelte } from '@threlte/core'
-import { session } from '$lib/stores'
+import { session } from '$lib/internal/stores'
 import { onDestroy } from 'svelte'
 
 const quaternion = new THREE.Quaternion()
