+++ src/lib/index.ts
@@ -1,12 +1,12 @@
-export { default as ARButton } from './button/button-ar.svelte'
-export { default as VRButton } from './button/button-vr.svelte'
-export { default as XRButton } from './button/button-xr.svelte'
-export { default as Controller } from './controller/controller.svelte'
-export { default as Controllers } from './controller/controllers.svelte'
-export { default as Hand } from './controller/hand.svelte'
-export { default as Hands } from './controller/hands.svelte'
-export { default as TeleportControls } from './teleport/teleport.svelte'
-export { default as XR } from './xr.svelte'
+export { default as ARButton } from './components/button-ar.svelte'
+export { default as VRButton } from './components/button-vr.svelte'
+export { default as XRButton } from './components/button-xr.svelte'
+export { default as Controller } from './components/controller.svelte'
+export { default as Controllers } from './components/controllers.svelte'
+export { default as Hand } from './components/hand.svelte'
+export { default as Hands } from './components/hands.svelte'
+export { default as TeleportControls } from './components/teleport.svelte'
+export { default as XR } from './components/xr.svelte'
 export { getXRSupportState } from './lib/get-xr-support-state'
 export { toggleXRSession } from './lib/toggle-xr-session'
 export { handJoints } from './lib/hand-joints'
